Added CMYK file-check & fixed .JPG extension bug.
Added a if-condition to check if a uploaded JPEG image is CMYK. If it is, echo error and die(). The tool currently only supports RGB JPEG's.
It also can handle CMYK PNG's. 

When a JPEG had a uppercase .JPG extension, the DPI wouldn't change. Now added strtolower and fixed it.